+++ src/main/java/com/personal/sistemaPersonal/controller/TreinoController.java
@@ -35,7 +35,7 @@ public class TreinoController {
     @RequestMapping("/showForm/{idAluno}")
     public String showForm(@PathVariable("idAluno") Integer idAluno, Model model) {
         Treino treino = new Treino();
-        treino.setFichaTreino(alunoService.getAlunoById(idAluno).getFicha_treino());
+        treino.setFicha_treino(alunoService.getAlunoById(idAluno).getFicha_treino());
         treino.setData_criacao(LocalDate.now());
 
         model.addAttribute("treino", treino);
@@ -55,7 +55,7 @@ public class TreinoController {
     @RequestMapping("/addTreino")
     public String addTreino(@ModelAttribute("treino") Treino treino) {
         treinoService.save(treino);
-        FichaTreino fichaTreino = fichaTreinoService.findById(treino.getFicha_treino().getIdFichaTreino());
+        FichaTreino fichaTreino = fichaTreinoService.findById(treino.getFicha_treino().getId());
 
         return "redirect:/aluno/detailsAluno/" + fichaTreino.getAluno().getId();
     }
